Add GetStub<T> method
a. GetStub<T> - This is to improve the readability of Unit Tests as
developers often use mocks for stub etc. The new method would provide
the same mocked instance but it can be used as a stub to setup data with
the test. More info
http://martinfowler.com/articles/mocksArentStubs.html

b. Add addition API comments to make the method more meaningful/extra
description.